+++ app/models/shoppe/product.rb
@@ -16,7 +16,7 @@ module Shoppe
     belongs_to :tax_rate, :class_name => "Shoppe::TaxRate"
     has_many :order_items, :dependent => :restrict_with_exception, :class_name => 'Shoppe::OrderItem', :as => :ordered_item
     has_many :orders, :through => :order_items, :class_name => 'Shoppe::Order'
-    has_many :stock_level_adjustments, :dependent => :destroy, :class_name => 'Shoppe::StockLevelAdjustment'
+    has_many :stock_level_adjustments, :dependent => :destroy, :class_name => 'Shoppe::StockLevelAdjustment', :as => :item
   
     # Validations
     validates :product_category_id, :presence => true