+++ db/migrate/20131022093538_stock_level_adjustments_should_be_polymorphic.rb
@@ -1,0 +1,6 @@
+class StockLevelAdjustmentsShouldBePolymorphic < ActiveRecord::Migration
+  def change
+    rename_column :shoppe_stock_level_adjustments, :product_id, :item_id
+    add_column :shoppe_stock_level_adjustments, :item_type, :string, :after => :item_id
+  end
+end