+++ test/dummy/db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20131022092904) do
+ActiveRecord::Schema.define(version: 20131022093538) do
 
   create_table "nifty_attachments", force: true do |t|
     t.integer  "parent_id"
@@ -157,13 +157,14 @@ ActiveRecord::Schema.define(version: 20131022092904) do
   end
 
   create_table "shoppe_stock_level_adjustments", force: true do |t|
-    t.integer  "product_id"
+    t.integer  "item_id"
     t.string   "description"
     t.integer  "adjustment",  default: 0
     t.string   "parent_type"
     t.integer  "parent_id"
     t.datetime "created_at"
     t.datetime "updated_at"
+    t.string   "item_type"
   end
 
   create_table "shoppe_tax_rates", force: true do |t|